5/13 add timer to GameRun

+++ Project2_Console_Mode/Project2_Console_Mode/GameManager.cpp
@@ -66,7 +66,7 @@ void File::Input(string& data, int& color, string& character, int& x1, int& y1, 
 //output to a file
 void File::Output() {
 	fstream out(filename, ios::out | ios::trunc);
-	for (int i = 0; i < gameRecord.size(); i++)
+	for (size_t i = 0; i < gameRecord.size(); i++)
 		out << gameRecord[i] << '\n';
 	out.close();
 }
@@ -80,8 +80,8 @@ Board::Board() {
 }
 
 void Board::initialization() {
-	for (int i = 0; i < 10; i++) {
-		for (int j = 0; j < 9; j++) {
+	for (size_t i = 0; i < 10; i++) {
+		for (size_t j = 0; j < 9; j++) {
 			board[i][j] = Null(j, i);
 		}
 	}
@@ -181,15 +181,14 @@ void Board::moveChess(File& file, int& color, string& character, int& x1, int& y
 
 void Board::showPath(vector<Pos>& cango) {
 	gotoxy(0, 20);
-	for (int i = 0; i < cango.size(); i++) {
-		
+	for (size_t i = 0; i < cango.size(); i++) {
 		cout << '(' << cango[i].x << ',' << cango[i].y << ')' << ' ';
 	}
 	return;
 }
 
 bool Board::checkcango(int& x2, int& y2, vector<Pos>& cango) {
-	for (int i = 0; i < cango.size(); i++) {
+	for (size_t i = 0; i < cango.size(); i++) {
 		if (x2 == cango[i].x && y2 == cango[i].y) return true;
 	}
 	return false;
